+++ src/keybinds/config.rs
@@ -75,6 +75,7 @@ pub struct LogTabKeybindsConfig {
     pub search: Option<Keybind>,
     pub set_bookmark: Option<Keybind>,
     pub open_files: Option<Keybind>,
+    pub open_tree: Option<Keybind>,
     pub copy_change_id: Option<Keybind>,
     pub copy_rev: Option<Keybind>,
     pub rebase: Option<Keybind>,